Added SQL and more styles + refactorings

+++ Rapport/Rapport/Pexels/IPexelsRestClient.cs
@@ -1,12 +1,12 @@
 ﻿using System.Threading.Tasks;
 using JetBrains.Annotations;
-using Rapport.Data.DTO.Pexels.Rapport.DTO;
+using Rapport.Data.DTO.Pexels;
 using RestSharp;
 
 namespace Rapport.Pexels
 {
     public interface IPexelsRestClient
     {
-        Task<IRestResponse<SearchResponse>> SearchAsync([NotNull] string query, byte resultsPerPage = 15, int startPage = 1);
+        Task<IRestResponse<SearchResponse>> SearchAsync([NotNull] string query, int resultsPerPage = 15, int startPage = 1);
     }
 }
